+++ doc/commands/branches.md
@@ -1,0 +1,10 @@
+# Branches
+
+A branch:
+- is a ref under ref/heads
+
+# commands that creates branches
+
+`git branch -c`
+`git checkout -b`
+`git switch -c`
